refactor: remove shallow node package
